refactor(hook): extract context-hook formatters to keep the file under the large-file bar
The line-aware change pushed context-hook.ts to 519 lines — over dxkit's own
500-line large-file threshold, which its self-guardrail (correctly) flagged.
Rather than allowlist our own file, practice the rule: split the three pure
presentation functions (formatFileContext, formatFileLineContext,
formatHookContext) into context-hook-format.ts. context-hook.ts (338 lines)
keeps the stdin/route/dedup orchestration; the new module (197 lines) is pure,
trivially testable string-building. No behavior change.
